backend/Himalaya/requirements.txt
asgiref==3.2.10
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
dataclasses==0.7
Django==3.0.5
django-cors-headers==3.4.0
django-filter==2.3.0
django-mysql==3.7.1
django-phonenumber-field==4.0.0
django-twilio==0.13.0.post1
djangorestframework==3.11.0
djongo==1.3.3
dnspython==2.0.0
fpdf==1.7.2
html5lib==1.1
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mysqlclient==2.0.0
pdfkit==0.6.1
phonenumbers==8.12.7
Pillow==7.2.0
PyJWT==1.7.1
pymongo==3.10.1
PyPDF2==1.26.0
pytz==2020.1
reportlab==3.5.46
requests==2.24.0
schedule==0.6.0
six==1.15.0
sqlparse==0.2.4
twilio==6.44.2
urllib3==1.25.10
webencodings==0.5.1
xhtml2pdf==0.2.4
